# docs/swagger.yaml
basePath: /v1
definitions:
...
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.LoginResponse:
    properties:
      success:
        type: boolean
    type: object
  auth.RegisterRequest:
    properties:
      birthday:
        type: string
      email:
        type: string
      firstName:
        type: string
      gender:
        allOf:
        - $ref: '#/definitions/users.Gender'
        enum:
        - male
        - female
      lastName:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - birthday
    - email
    - firstName
    - gender
    - lastName
    - password
    type: object
  auth.RegisterResponse:
    properties:
      success:
        type: boolean
    type: object
...
  users.Gender:
    enum:
    - male
    - female
    type: string
    x-enum-varnames:
    - Male
    - Female
  users.GetUserResponse:
    properties:
      birthday:
        type: string
      createdAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      gender:
        $ref: '#/definitions/users.Gender'
      id:
        type: integer
      lastName:
        type: string
      role:
        $ref: '#/definitions/users.Role'
      status:
        $ref: '#/definitions/users.Status'
      updatedAt:
        type: string
    type: object
  users.Role:
    enum:
    - user
    - admin
    - moderator
    type: string
    x-enum-varnames:
    - UserRole
    - AdminRole
    - ModeratorRole
  users.Status:
    enum:
    - active
    - banned
    - pending
    - deleted
    type: string
    x-enum-varnames:
    - Active
    - Banned
    - Pending
    - Deleted
...
host: localhost:777
info:
  contact: {}
  description: This is a Concert booking application
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Concert booking API
  version: "1.0"
paths:
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: 'Login user, set token in cookie and return success: true'
      parameters:
      - description: LoginRequest credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged in
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Login a user
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user and return authentication token
      parameters:
      - description: Registration credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully registered
          schema:
            $ref: '#/definitions/auth.RegisterResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Register a new user
      tags:
      - auth
  /api/v1/users/{id}:
    get:
      description: Returns a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.GetUserResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get user by ID
      tags:
      - Users
...
swagger: "2.0"